moved and renamed test utility helpers file
